feat: refactor wallet balance fetching and address formatting into utility functions; update NavBarResponsive and WalletDetails components to use shared logic
