+++ abroad-ui/src/index.css
@@ -1,82 +1,2 @@
 @import "tailwindcss";
 @import "./fonts.css";
-
-/* Custom scrollbar styles for WalletDetails */
-.wallet-details-scroll {
-  scrollbar-width: thin;
-  scrollbar-color: transparent transparent;
-  transition: scrollbar-color 0.3s ease;
-}
-
-.wallet-details-scroll:hover {
-  scrollbar-color: #d1d5db transparent;
-}
-
-.wallet-details-scroll::-webkit-scrollbar {
-  width: 6px;
-  position: absolute;
-  right: 0;
-}
-
-.wallet-details-scroll::-webkit-scrollbar-track {
-  background: transparent;
-}
-
-.wallet-details-scroll::-webkit-scrollbar-thumb {
-  background-color: transparent;
-  border-radius: 10px;
-  border: transparent;
-  transition: background-color 0.3s ease;
-}
-
-.wallet-details-scroll:hover::-webkit-scrollbar-thumb {
-  background-color: #d1d5db;
-}
-
-.wallet-details-scroll::-webkit-scrollbar-thumb:hover {
-  background-color: #9ca3af;
-}
-
-.wallet-details-scroll::-webkit-scrollbar-thumb:active {
-  background-color: #6b7280;
-}
-
-/* Hide scrollbar corner buttons/triangles */
-.wallet-details-scroll::-webkit-scrollbar-button {
-  display: none !important;
-  width: 0 !important;
-  height: 0 !important;
-  background: transparent !important;
-}
-
-.wallet-details-scroll::-webkit-scrollbar-button:start:decrement,
-.wallet-details-scroll::-webkit-scrollbar-button:end:increment {
-  display: none !important;
-  width: 0 !important;
-  height: 0 !important;
-  background: transparent !important;
-}
-
-.wallet-details-scroll::-webkit-scrollbar-button:single-button {
-  display: none !important;
-  width: 0 !important;
-  height: 0 !important;
-}
-
-.wallet-details-scroll::-webkit-scrollbar-button:double-button {
-  display: none !important;
-  width: 0 !important;
-  height: 0 !important;
-}
-
-.wallet-details-scroll::-webkit-scrollbar-button:vertical:start:increment,
-.wallet-details-scroll::-webkit-scrollbar-button:vertical:end:decrement {
-  display: none !important;
-  width: 0 !important;
-  height: 0 !important;
-}
-
-.wallet-details-scroll::-webkit-scrollbar-corner {
-  background: transparent !important;
-  display: none !important;
-}
